chore: trim GUI
