+++ GUI/components/__init__.py
@@ -1,0 +1,1 @@
+from GUI.components.menu import ToolMenu
